refactor: board service test 분리
